Made some changes in order to make Jake's code more readable. I mostly changed variable names.
P.S I added a getHeading() method to all "Direction"s but did not change
the code in Task.java to use it. I didn't check the semantics of your
code for possible ramifications but you (Jake) should it's a much better
way of tracking the heading.  Moreover you should consider making North
0;